Revert "Revert "Webhooks to support event type association"" (#2033)
* Revert "Revert "Webhooks to support event type association (#1889)" (#2032)"

This reverts commit 71e74b8320301df770f4035d72b19e0035e42fcf.

* Fixes turbo DB deploy

* Update turbo.json
