+++ apps/web/lib/webhooks/subscriptions.tsx
@@ -2,13 +2,27 @@ import { WebhookTriggerEvents } from "@prisma/client";
 
 import prisma from "@lib/prisma";
 
-const getSubscribers = async (userId: number, triggerEvent: WebhookTriggerEvents) => {
+export type GetSubscriberOptions = {
+  userId: number;
+  eventTypeId: number;
+  triggerEvent: WebhookTriggerEvents;
+};
+
+const getSubscribers = async (options: GetSubscriberOptions) => {
+  const { userId, eventTypeId } = options;
   const allWebhooks = await prisma.webhook.findMany({
     where: {
-      userId: userId,
+      OR: [
+        {
+          userId,
+        },
+        {
+          eventTypeId,
+        },
+      ],
       AND: {
         eventTriggers: {
-          has: triggerEvent,
+          has: options.triggerEvent,
         },
         active: {
           equals: true,
