+++ apps/web/public/static/locales/en/common.json
@@ -314,6 +314,7 @@
   "create_new_webhook_to_account": "Create a new webhook to your account",
   "new_webhook": "New Webhook",
   "receive_cal_meeting_data": "Receive Cal meeting data at a specified URL, in real-time, when an event is scheduled or cancelled.",
+  "receive_cal_event_meeting_data": "Receive Cal meeting data at a specified URL, in real-time, when this event is scheduled or cancelled.",
   "responsive_fullscreen_iframe": "Responsive full screen iframe",
   "loading": "Loading...",
   "standard_iframe": "Standard iframe",
